+++ changelog.md
@@ -5,6 +5,7 @@
 ### Fixes
 
 - [#4864](https://github.com/ignite/cli/pull/4864) Mismatch for message names.
+- [#4735](https://github.com/ignite/cli/issues/4735) Cleanup `xgenny` runner to avoid duplicated generators.
 
 ## [`v29.7.0`](https://github.com/ignite/cli/releases/tag/v29.7.0)